Connect

+++ client/src/entities/TaskCard/components/RemoveTaskForm.tsx
@@ -10,9 +10,10 @@ import { useDispatch, useSelector } from "react-redux";
 
 import { useState } from "react";
 import { useList } from "@/app/list-provider/list-provider";
-import { addRemoveActivity } from "@/app/store/activity-slice/activity-slice";
 import { AppDispatch, RootState } from "@/app/store/store";
 import { fetchDeleteTodo } from "@/app/store/todo-slice/thunks/fetch-delete-todo";
+import { RemoveActivity } from "@/app/store/activity-slice/activity-slice";
+import { fetchAddActivity } from "@/app/store/activity-slice/thunks/fetch-add-activity";
 
 export default function RemoveTaskForm({
   children,
@@ -55,14 +56,25 @@ export default function RemoveTaskForm({
           <Button
             className="w-1/2"
             onClick={() => {
+              const removeAcitityPayload: RemoveActivity = {
+                date: new Date(),
+                taskId,
+                taskName: name,
+                targetList: listId,
+                type: "REMOVE",
+              };
+
+              const ownerId = localStorage.getItem("token");
+
+              if (!ownerId) return;
+
               dispatch(
-                addRemoveActivity({
-                  taskId,
-                  targetList: listId,
-                  type: "remove",
-                  taskName: name,
+                fetchAddActivity({
+                  activityData: removeAcitityPayload,
+                  ownerId,
                 })
               );
+
               dispatch(fetchDeleteTodo({ id: taskId }));
               setIsOpen(false);
             }}
